Extract pushups function into separate module
